+++ crud_basic/src/main/java/com/sena/crud_basic/repository/IBookingDetailRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sena.crud_basic.model.BookingDetail;
 
-public interface IBookingDetailRepository extends JpaRepository<BookingDetail, Long> {
+public interface IBookingDetailRepository extends JpaRepository<BookingDetail, Integer> {
 
 }
