+++ crud_basic/src/main/java/com/sena/crud_basic/repository/IBookingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sena.crud_basic.model.Booking;
 
-public interface IBookingRepository extends JpaRepository<Booking, Long>{
+public interface IBookingRepository extends JpaRepository<Booking, Integer>{
 
 }
