+++ crud_basic/src/main/java/com/sena/crud_basic/repository/IEmployeeHotelRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sena.crud_basic.model.EmployeeHotel;
 
-public interface IEmployeeHotelRepository extends JpaRepository<EmployeeHotel, Long> {
+public interface IEmployeeHotelRepository extends JpaRepository<EmployeeHotel, Integer> {
 
 }
