+++ crud_basic/src/main/java/com/sena/crud_basic/repository/IEmployeeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sena.crud_basic.model.Employee;
 
-public interface IEmployeeRepository extends JpaRepository<Employee, Long>{
+public interface IEmployeeRepository extends JpaRepository<Employee, Integer>{
 
 }
