+++ crud_basic/src/main/java/com/sena/crud_basic/repository/IHotelRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sena.crud_basic.model.Hotel;
 
-public interface IHotelRepository extends JpaRepository<Hotel, Long> {
+public interface IHotelRepository extends JpaRepository<Hotel, Integer> {
 
 }
