+++ crud_basic/src/main/java/com/sena/crud_basic/repository/IInvoiceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sena.crud_basic.model.Invoice;
 
-public interface IInvoiceRepository extends JpaRepository<Invoice, Long>{
+public interface IInvoiceRepository extends JpaRepository<Invoice, Integer>{
 
 }
