+++ crud_basic/src/main/java/com/sena/crud_basic/repository/IServiceRepository.java
@@ -1,9 +1,8 @@
 package com.sena.crud_basic.repository;
-
-import java.security.Provider.Service;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IServiceRepository extends JpaRepository<Service, Long> {
+import com.sena.crud_basic.model.Service;
+
+public interface IServiceRepository extends JpaRepository<Service, Integer> {
 
 }
